modify request for jwt user
