+++ src/server/handler_authenticate.go
@@ -2,9 +2,7 @@ package server
 
 import (
 	"strings"
-	"time"
 
-	"github.com/dgrijalva/jwt-go"
 	response "github.com/linkernetworks/vortex/src/net/http"
 	"github.com/linkernetworks/vortex/src/web"
 )
@@ -15,7 +13,7 @@ type UserCredentials struct {
 	Password string `json:"password"`
 }
 
-func authenticateHandler(ctx *web.Context) {
+func loginHandler(ctx *web.Context) {
 	_, req, resp := ctx.ServiceProvider, ctx.Request, ctx.Response
 	user := UserCredentials{}
 	if err := req.ReadEntity(&user); err != nil {
@@ -39,17 +37,4 @@ func authenticateHandler(ctx *web.Context) {
 		Error:   false,
 		Message: tokenString,
 	})
-}
-
-func generateToken(userUUID string) (string, error) {
-	token := jwt.New(jwt.SigningMethodHS256)
-	token.Claims = jwt.MapClaims{
-		// issuer of the claim
-		"exp": time.Now().Add(time.Hour * time.Duration(1)).Unix(),
-		// issued-at time
-		"iat": time.Now().Unix(),
-		// the subject of this token. This is the user associated with the relevant action
-		"sub": userUUID,
-	}
-	return token.SignedString([]byte(SecretKey))
 }
